+++ src/torch/cpu/ops_impl_cpu.h
@@ -24,7 +24,7 @@ void range_impl(Tensor a, double start, double end, double step);
 
 void fill_impl(Tensor& a, double value);
 void fill_impl(Tensor& a, Tensor value);
-void fill_impl(Tensor& a, Tensor values, int dim);
+void fill_impl(Tensor& a, Tensor values, int64_t dim);
 void permute_impl(Tensor& src, Tensor result, SizeType new_dims);
 
 void uniform_impl(Tensor& t, double mi, double ma);
